working zooming and move-zooming on center cam position
- starting camera follow now (with bounds)

+++ nova2d/nova-zero/src/graphics/Image.h
@@ -36,8 +36,8 @@ namespace novazero
 			void OriginCenter() { GetDrawable()->OriginCenter(); }
 			void OriginTopLeft() { GetDrawable()->OriginTopLeft(); }
 
-			int OffsetX() { return GetDrawable()->OffsetX(); }
-			int OffsetY() { return GetDrawable()->OffsetY(); }
+			float OffsetX() { return GetDrawable()->OffsetX(); }
+			float OffsetY() { return GetDrawable()->OffsetY(); }
 
 			bool operator==(const Image& other);
